Update Developer Guide and refactor storage file
