+++ src/main/java/seedu/duke/controllers/MainController.java
@@ -1,6 +1,6 @@
 package seedu.duke.controllers;
 
-import seedu.duke.models.schema.Storage;
+import seedu.duke.storage.StorageManager;
 import seedu.duke.models.schema.Student;
 import seedu.duke.models.schema.CommandManager;
 import seedu.duke.models.schema.UserCommand;
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 
 import static seedu.duke.controllers.ModuleServiceController.validateMajorInput;
 
-import static seedu.duke.models.schema.Storage.saveTimetable;
+import static seedu.duke.storage.StorageManager.saveTimetable;
 import static seedu.duke.utils.Utility.detectInternet;
 import static seedu.duke.utils.Utility.saveStudentData;
 import static seedu.duke.views.Ui.displayWelcome;
@@ -27,14 +27,14 @@ import static seedu.duke.views.CommandLineView.displayGetYear;
 import static seedu.duke.views.Ui.showLoadingAnimation;
 import static seedu.duke.views.Ui.stopLoadingAnimation;
 
-import static seedu.duke.models.schema.Storage.saveSchedule;
+import static seedu.duke.storage.StorageManager.saveSchedule;
 
 
 public class MainController {
     private final Parser parser;
     private final Student student;
     private final CommandManager commandManager;
-    private Storage storage;
+    private StorageManager storageManager;
 
     private Ui ui;
 
@@ -62,7 +62,7 @@ public class MainController {
         initialiseUser();
         displayReady();
         handleUserInputTillExitCommand();
-        saveStudentData(storage,student);
+        saveStudentData(storageManager,student);
         displayGoodbye();
     }
 
@@ -76,23 +76,23 @@ public class MainController {
      */
     public void initialiseUser() throws IOException {
 
-        storage = new Storage();
+        storageManager = new StorageManager();
         try {
             System.out.println("Attempting to retrieve data from save file... Sorry this takes a while!");
             showLoadingAnimation();
             // Load name, major and year from studentDetails.txt file
-            ArrayList<String> studentDetails = storage.loadStudentDetails();
+            ArrayList<String> studentDetails = storageManager.loadStudentDetails();
 
             // Set name, major and year from loaded data, throws exception if file is corrupted.
             setStudentDetails(studentDetails);
 
             // Load and set schedule from schedule.txt file
-            student.setSchedule(storage.loadSchedule());
+            student.setSchedule(storageManager.loadSchedule());
 
             // Load timetable from timetable.txt file
             try {
                 student.updateTimetable();
-                storage.addEventsToStudentTimetable(storage.loadTimetable(student), student);
+                storageManager.addEventsToStudentTimetable(storageManager.loadTimetable(student), student);
 
             } catch (TimetableUnavailableException e) {
                 // no modules in current sem, do nothing
@@ -135,7 +135,7 @@ public class MainController {
     }
 
     public void resetStorageData() throws IOException {
-        storage.createUserStorageFile();
+        storageManager.createUserStorageFile();
 
         String userInput;
 
@@ -158,7 +158,7 @@ public class MainController {
             userInput = ui.getUserCommand("Please enter your current academic year: ").trim();
         } while (!Parser.isValidAcademicYear(userInput.toUpperCase()));
         student.setYear(userInput.toUpperCase());
-        storage.saveStudentDetails(student);
+        storageManager.saveStudentDetails(student);
 
         //get blank schedule.txt
         student.setSchedule(new Schedule());
