+++ src/main/java/seedu/duke/storage/FileStorage.java
@@ -1,0 +1,4 @@
+package seedu.duke.storage;
+
+public class FileStorage {
+}
