+++ src/test/java/seedu/duke/models/logic/DataRepositoryTest.java
@@ -1,7 +1,7 @@
 package seedu.duke.models.logic;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.models.schema.Storage;
+import seedu.duke.storage.StorageManager;
 
 import java.util.ArrayList;
 
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 class DataRepositoryTest {
     @Test
     void validRequirementsReturned() {
-        ArrayList<String> cegRequirementArray = Storage.getRequirements("CEG");
+        ArrayList<String> cegRequirementArray = StorageManager.getRequirements("CEG");
         int numberOfRequiredCegMods = 31;
         assertEquals(numberOfRequiredCegMods,cegRequirementArray.size());
     }
